auth/server/requirements.txt
paho-mqtt==1.6.1
pip==22.0.4
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
setuptools==57.0.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
wheel==0.36.2
